+++ scripts/createSprite.js
@@ -1,3 +1,0 @@
-const { createSprite } = require('create-svg-icon-sprite');
-
-createSprite();